Changing offsetVectorX and offsetVectorY to float from double.
git-svn-id: http://chianti.ucsd.edu/svn@3587 0ecc0d97-ab19-0410-9704-bfe1a75892f5

+++ csplugins/trunk/ucsd/nlandys/fung/src/cytoscape/fung/Label.java
@@ -30,12 +30,12 @@ public abstract class Label
   final double m_scaleFactor;
   final Paint m_paint;
   final byte m_textAnchor;
-  final double m_offsetVectorX;
-  final double m_offsetVectorY;
+  final float m_offsetVectorX;
+  final float m_offsetVectorY;
   final byte m_justify;
 
   Label(String text, Font font, double scaleFactor, Paint paint,
-        byte textAnchor, double offsetVectorX, double offsetVectorY,
+        byte textAnchor, float offsetVectorX, float offsetVectorY,
         byte justify)
   {
     m_text = text;
